Moved logger to avoid circular import
