+++ clinicadl/utils/maps_manager/maps_manager.py
@@ -12,7 +12,6 @@ import pandas as pd
 import torch
 from torch.utils.data import DataLoader
 
-from clinicadl.cmdline import setup_logging
 from clinicadl.utils.caps_dataset.data import (
     get_transforms,
     load_data_test,
@@ -26,6 +25,7 @@ from clinicadl.utils.exceptions import (
     ClinicaDLDataLeakageError,
     MAPSError,
 )
+from clinicadl.utils.logger import setup_logging
 from clinicadl.utils.maps_manager.logwriter import LogWriter
 from clinicadl.utils.maps_manager.maps_manager_utils import (
     add_default_values,
